Add files via upload
readepub.tmp extracts to /tmp/readepub instead of "requiring" you to select a path! This is probably the one I'd focus on if it were something I'd need to expand upon. The truth of it is that this will just be something I use to conveniently open epubs without having to do it manually each time.